fix(api): retain remote cover URLs when no local poster resolves
The cover resolution in library, search, and playlist routes
unconditionally overwrote manga.information.cover_url with the result
of poster_exists().and_then(path_to_file_url), clearing an existing
remote cover_url to None for any manga without a downloaded local
poster. This caused non-base library, search, and playlist views to
lose their covers.

Extract a shared resolve_manga_covers helper that only overwrites
cover_url when a local poster actually resolves to a file:// URL,
preserving the existing remote URL otherwise.
